requirements.txt
anytree==2.8.0
astroid==2.3.3
beautifulsoup4==4.8.2
boto3==1.11.9
botocore==1.14.9
bullet==2.1.0
certifi==2019.11.28
chardet==3.0.4
cliff==2.18.0
cmd2==0.8.9
configobj==5.0.6
configparser==4.0.2
confuse==1.0.0
docutils==0.15.2
filelock==3.0.12
future==0.18.2
gnupg==2.3.1
humanfriendly==4.18
idna==2.8
importlib-metadata==1.4.0
isort==4.3.21
Jinja2==2.10.3
jmespath==0.9.4
lazy-object-proxy==1.4.3
lxml==4.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.1
oauthlib==3.1.0
packaging==20.1
pbr==5.4.4
pexpect==4.8.0
Pillow==7.0.0
pluggy==0.13.1
prettytable==0.7.2
property-manager==2.3.1
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.1
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.6
pyperclip==1.7.0
python-dateutil==2.8.1
python-dotenv==0.10.5
python-twitter==3.5
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.3.2
six==1.14.0
soupsieve==1.9.5
stevedore==1.31.0
toml==0.10.0
tox==3.14.3
typed-ast==1.4.1
urllib3==1.25.8
verboselogs==1.7
virtualenv==16.7.9
wcwidth==0.1.8
wrapt==1.11.2
xkcdpass==1.17.3
yamlreader==3.0.4
zipp==2.1.0
